Bot/requirements.txt
python-telegram-bot
requests
python-dotenv
gpt4all